% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{AffineTransform}
\alias{AffineTransform}
\title{AffineTransform Object}
\usage{
AffineTransform(
  scaleY = NULL,
  translateY = NULL,
  translateX = NULL,
  scaleX = NULL,
  shearX = NULL,
  unit = NULL,
  shearY = NULL
)
}
\arguments{
\item{scaleY}{The Y coordinate scaling element}

\item{translateY}{The Y coordinate translation element}

\item{translateX}{The X coordinate translation element}

\item{scaleX}{The X coordinate scaling element}

\item{shearX}{The X coordinate shearing element}

\item{unit}{The units for translate elements}

\item{shearY}{The Y coordinate shearing element}
}
\value{
AffineTransform object
}
\description{
AffineTransform Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
AffineTransform uses a 3x3 matrix with an implied last row of \link{ 0 0 1 } to transform source coordinates (x,y) into destination coordinates (x', y') according to: x' x = shear_y scale_y translate_y 1 \link{ 1 } After transformation, x' = scale_x * x + shear_x * y + translate_x; y' = scale_y * y + shear_y * x + translate_y; This message is therefore composed of these six matrix elements.
}
\concept{AffineTransform functions}
