% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{CreateShapeRequest}
\alias{CreateShapeRequest}
\title{CreateShapeRequest Object}
\usage{
CreateShapeRequest(shapeType = NULL, elementProperties = NULL, objectId = NULL)
}
\arguments{
\item{shapeType}{The shape type}

\item{elementProperties}{The element properties for the shape}

\item{objectId}{A user-supplied object ID}
}
\value{
CreateShapeRequest object
}
\description{
CreateShapeRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Creates a new shape.
}
\concept{CreateShapeRequest functions}
