% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_functions.R
\name{presentations.pages.getThumbnail}
\alias{presentations.pages.getThumbnail}
\title{Generates a thumbnail of the latest version of the specified page in the presentation and returns a URL to the thumbnail image. This request counts as an \href{/slides/limits}{expensive read request} for quota purposes.}
\usage{
presentations.pages.getThumbnail(
  presentationId,
  pageObjectId,
  thumbnailProperties.thumbnailSize = NULL,
  thumbnailProperties.mimeType = NULL
)
}
\arguments{
\item{presentationId}{The ID of the presentation to retrieve}

\item{pageObjectId}{The object ID of the page whose thumbnail to retrieve}

\item{thumbnailProperties.thumbnailSize}{The optional thumbnail image size}

\item{thumbnailProperties.mimeType}{The optional mime type of the thumbnail image}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/drive
\item https://www.googleapis.com/auth/drive.file
\item https://www.googleapis.com/auth/drive.readonly
\item https://www.googleapis.com/auth/presentations
\item https://www.googleapis.com/auth/presentations.readonly
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/drive, https://www.googleapis.com/auth/drive.file, https://www.googleapis.com/auth/drive.readonly, https://www.googleapis.com/auth/presentations, https://www.googleapis.com/auth/presentations.readonly)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/slides/}{Google Documentation}
}
