% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_functions.R
\docType{package}
\name{slides_googleAuthR}
\alias{slides_googleAuthR}
\title{Google Slides API
Reads and writes Google Slides presentations.}
\description{
Auto-generated code by googleAuthR::gar_create_api_skeleton
at 2023-02-11 20:37:07
filename: C:/Users/isaac/Documents/slides_api/googleslidesv1.auto/R/slides_functions.R
api_json: api_json
}
\details{
Authentication scopes used are:
\itemize{
\item https://www.googleapis.com/auth/drive
\item https://www.googleapis.com/auth/presentations
\item https://www.googleapis.com/auth/spreadsheets.readonly
\item https://www.googleapis.com/auth/spreadsheets
\item https://www.googleapis.com/auth/presentations.readonly
\item https://www.googleapis.com/auth/drive.file
\item https://www.googleapis.com/auth/drive.readonly
}
}
