% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{UpdateParagraphStyleRequest}
\alias{UpdateParagraphStyleRequest}
\title{Google Slides API Objects
Reads and writes Google Slides presentations.}
\usage{
UpdateParagraphStyleRequest(
  cellLocation = NULL,
  textRange = NULL,
  objectId = NULL,
  fields = NULL,
  style = NULL
)
}
\arguments{
\item{cellLocation}{The location of the cell in the table containing the paragraph(s) to style}

\item{textRange}{The range of text containing the paragraph(s) to style}

\item{objectId}{The object ID of the shape or table with the text to be styled}

\item{fields}{The fields that should be updated}

\item{style}{The paragraph's style}
}
\value{
UpdateParagraphStyleRequest object
}
\description{
Auto-generated code by googleAuthR::gar_create_api_objects
at 2023-02-11 20:37:07
filename: C:/Users/isaac/Documents/slides_api/googleslidesv1.auto/R/slides_objects.R
api_json: api_json
}
\details{
Objects for use by the functions created by googleAuthR::gar_create_api_skeleton
UpdateParagraphStyleRequest Object

Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Updates the styling for all of the paragraphs within a Shape or Table that overlap with the given text index range.
}
\concept{UpdateParagraphStyleRequest functions}
